requirements.txt
hermes_python>=0.1
paho-mqtt
